Focus scope on notion of unitless Amount
In discussions among various TC39 delegates, it has become
clear that the concept of a unit brings up a number of
challenges that can be cleanly separated from the more
fundamental problem of representing mathematical values
together with a precision. This is an important part of the
notion of measurement that is a subset of the original
motivation and argument for Measure. We propose to separate
these challenges from those that come from supporting
measurements, such as:

* the need for unit conversion,
* compound units (e.g., "foot-and-inch"),
* whether: some units (and which ones?) should be
  included (SI?) out-of-the-box,
* user-customizable units,
* arithmetic: involved in conversions (rationals?  decimals?
  binary floats?),

etc.